restore default NetworkPolicy + kubeapps
This is required until
https://github.com/controlplaneio-fluxcd/flux-operator/issues/51 can be
resolved in a release, if you don't set `networkPolicy` then you'll get
a wrong behavior from Flux, which should be deployed with NetworkPolicy
resources by default.

The `allow-from-kubeapps` policy is so that cozy-dashboard can poll and
represent the status of Helm artifacts, it is still needed when this fix
is in.

Signed-off-by: Kingdon Barrett <kingdon+github@tuesdaystudios.com>