Fine-tuning Flux configuration
Don't set the `interval` so short on HelmReleases, with this many
HelmReleases that really hamstrings the control plane.

Also, copy the install/upgrade remediation config from system packages
to the Kubernetes templates for addon packages (cilium, flux, ingress) -
in my testing the ingress-nginx chart fails every time the first time.
Maybe that should be filed as a separate issue, I haven't looked into
detail, it is some issue related to a secret not being created.

I assume it's a conflict with being installed at the same time as the
cert-manager addon (but I'm not sure and I can't look into it now!)

Signed-off-by: Kingdon Barrett <kingdon+github@tuesdaystudios.com>